requirements.txt
absl-py==2.4.0
accelerate==1.10.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==26.1.0
bitsandbytes==0.49.2
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.2.1
contourpy==1.3.0
cycler==0.12.1
datasets==4.5.0
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
exceptiongroup==1.3.1
faiss-cpu==1.13.0
fastapi==0.128.8
ffmpy==1.0.0
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
gradio==4.44.1
gradio_client==1.3.0
groq==1.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.2.1
nltk==3.9.4
numpy==2.0.2
orjson==3.11.5
packaging==26.1
pandas==2.3.3
peft==0.17.1
pillow==10.4.0
propcache==0.4.1
psutil==7.2.2
pyarrow==21.0.0
pydantic==2.13.2
pydantic_core==2.46.2
pydub==0.25.1
Pygments==2.20.0
pymongo==4.16.0
pyparsing==3.3.2
pypdf==6.10.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==15.0.0
rouge_score==0.1.2
ruff==0.15.11
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.12.0
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.39.0
websockets==12.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.1
